+++ CloudDemo_CSYE_6225/src/main/java/com/shreyas/CloudDemo/service/interfaces/HealthCheckService.java
@@ -1,4 +1,4 @@
-package com.shreyas.CloudDemo.service;
+package com.shreyas.CloudDemo.service.interfaces;
 
 public interface HealthCheckService {
     boolean isDataConnectionAvailable() throws Exception;
