+++ CloudDemo_CSYE_6225/src/test/java/com/shreyas/CloudDemo/controller/HealthCheckControllerTest.java
@@ -1,7 +1,7 @@
 package com.shreyas.CloudDemo.controller;
 
 import com.shreyas.CloudDemo.config.TestSecurityConfig;
-import com.shreyas.CloudDemo.service.HealthCheckService;
+import com.shreyas.CloudDemo.service.interfaces.HealthCheckService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
